chore: keep OpenCode provider files under size lint
